integers using 54 bits

+++ src/base/parser/grammar/ast.h
@@ -18,7 +18,7 @@ namespace ast
 {
     namespace x3 = boost::spirit::x3;
 
-    typedef boost::variant<std::string, int, float, bool> Value;
+    typedef boost::variant<std::string, int64_t, float, bool> Value;
 
     struct Element {
         std::string function; // If empty string then no function
